лабораторная 4/requirements.txt
fastapi[standard]
pydantic
sqlalchemy
bcrypt
passlib
pyjwt
python-jose